+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## master
 
+* Support :scope_class option to force a specific scope class, instead of using the class of the provided scope (jeremyevans)
 * Support fallback fixed locals for templates via :default_fixed_locals option (jeremyevans)
 * Support embedded fixed locals for templates via magic comments, if Tilt.extract_fixed_locals is set (jeremyevans)
 * Support fixed locals for templates via :fixed_locals option (jeremyevans)